DS-1815: fuse_table to table/fuse
