+++ query/src/datasources/table/fuse/io/mod.rs
@@ -14,11 +14,16 @@
 //
 
 // consider remove these, read_util seems to be enough (type could be inferred)
-pub(crate) mod segment_reader;
-pub(crate) mod snapshot_reader;
+mod segment_reader;
+mod snapshot_reader;
 // end
 
+mod block_appender;
+mod block_reader;
 mod reader_util;
 
-pub(crate) mod block_appender;
-pub(crate) mod block_reader;
+pub use block_appender::*;
+pub use block_reader::*;
+pub use reader_util::*;
+pub use segment_reader::*;
+pub use snapshot_reader::*;
