+++ query/src/datasources/table/fuse/io/snapshot_reader.rs
@@ -18,21 +18,23 @@ use std::sync::Arc;
 use common_exception::Result;
 
 use crate::datasources::dal::DataAccessor;
-use crate::datasources::fuse_table::io::reader_util::do_read_obj;
-use crate::datasources::fuse_table::io::reader_util::do_read_obj_async;
-use crate::datasources::fuse_table::meta::table_snapshot::SegmentInfo;
+use crate::datasources::table::fuse::do_read_obj;
+use crate::datasources::table::fuse::do_read_obj_async;
+use crate::datasources::table::fuse::TableSnapshot;
 use crate::sessions::DatafuseQueryContextRef;
 
-#[allow(dead_code)]
-pub fn read_segment(
+pub fn read_table_snapshot(
     da: Arc<dyn DataAccessor>,
     ctx: &DatafuseQueryContextRef,
     loc: &str,
-) -> Result<SegmentInfo> {
+) -> Result<TableSnapshot> {
     do_read_obj(da, ctx, loc)
 }
 
 #[allow(dead_code)]
-pub async fn read_segment_async(da: Arc<dyn DataAccessor>, loc: &str) -> Result<SegmentInfo> {
+pub async fn read_table_snapshot_async(
+    da: Arc<dyn DataAccessor>,
+    loc: &str,
+) -> Result<TableSnapshot> {
     do_read_obj_async(da, loc).await
 }
