+++ query/src/datasources/table/fuse/util/mod.rs
@@ -13,5 +13,14 @@
 //  limitations under the License.
 //
 
-pub mod meta_info_reader;
-pub mod table_snapshot;
+mod index_helpers;
+mod location_gen;
+mod projection_helper;
+mod statistic_helper;
+mod storage_scheme_helper;
+
+pub use index_helpers::range_filter;
+pub use location_gen::*;
+pub use projection_helper::project_col_idx;
+pub use statistic_helper::column_stats_reduce;
+pub use storage_scheme_helper::*;
