+++ query/src/datasources/table/fuse/util/projection_helper.rs
