+++ query/src/datasources/table/fuse/util/storage_scheme_helper.rs
